// lib/api/model/sectionDto.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface SectionDto {
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  id: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  courseOfferingId: number | string;
  name: string;
  createdAt: string;
}