// lib/api/model/attendanceMatrixResult.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { ModuleHeader } from './moduleHeader';
import type { StudentRow } from './studentRow';

export interface AttendanceMatrixResult {
  modules: ModuleHeader[];
  students: StudentRow[];
}